Add invalidPathException

+++ src/main/Utils/IOMethods.java
@@ -4,8 +4,14 @@ import java.io.*;
 import java.util.Scanner;
 
 import Encryption.CharAlgo.CharEncryptionAlgorithmAbstract;
+import Exceptions.invalidPathException;
 
 public class IOMethods {
+    private static void checkExistPath(String path){
+        if(!new File(path).exists())
+            throw new invalidPathException();
+    }
+
     /**
      * Scan the input file line by line and put the changed lines into the output line
      * @param inputPath path to input file
@@ -41,6 +47,8 @@ public class IOMethods {
      * @param line message to be written into file.
      */
     public static void writeLine(String path, String line) throws IOException {
+        if(!new File(path).exists())
+            throw new invalidPathException();
         Writer output = new BufferedWriter(new FileWriter(path, true));
         output.append(line);
         output.close();
@@ -51,6 +59,8 @@ public class IOMethods {
      * @param path file path and name
      */
     public static void createFile(String path) throws IOException {
+        if(!new File(path).exists())
+            throw new invalidPathException();
         final File myObj = new File(path);
         if(myObj.exists())
             if(!myObj.delete())
@@ -65,6 +75,8 @@ public class IOMethods {
      * @param message message to write into path
      */
     public static void writeToFile(String path, String message) throws IOException {
+        if(!new File(path).exists())
+            throw new invalidPathException();
         FileWriter myWriter = new FileWriter(path);
         myWriter.write(message);
         myWriter.close();
@@ -76,6 +88,8 @@ public class IOMethods {
      * @return string of the text to the given file combine and separated by \n char
      */
     public static String readFile(String path) throws FileNotFoundException {
+        if(!new File(path).exists())
+            throw new invalidPathException();
         StringBuilder txt = new StringBuilder();
         Scanner myReader = new Scanner(new File(path));
         while (myReader.hasNextLine())
@@ -90,6 +104,10 @@ public class IOMethods {
      * @param newPath file path to copy
      */
     public static void copyFile(String originalPath, String newPath) throws IOException {
+        if(!new File(originalPath).exists())
+            throw new invalidPathException();
+        if(!new File(newPath).exists())
+            throw new invalidPathException();
         File originalFile = new File(originalPath);
         File newFile = new File(newPath);
